Added a compat package. Moved jei inside.
Moved Jei to the compat package, started CT compat
