+++ build.gradle
@@ -28,6 +28,10 @@ repositories {
         name = "ModMaven"
         url = "https://modmaven.k-4u.nl"
     }
+    maven {
+        name "Crafttweaker Maven"
+        url "http://maven.blamejared.com/"
+    }
 }
 
 version = "0.1.1"
@@ -54,6 +58,8 @@ dependencies {
     compile fileTree(include: ["*.jar"], dir: "libs")
     deobfProvided "mezz.jei:jei_1.12.2:4.15.0.291"
     compile "mezz.jei:jei_1.12.2:4.15.0.291"
+    compile "CraftTweaker2:CraftTweaker2-MC1120-Main:1.12-4.1.20.554"
+    compile "com.blamejared:MTLib:3.0.4.8"
 }
 
 processResources
