+++ src/main/java/com/eerussianguy/firmalife/compat/crafttweaker/CTCracking.java
@@ -1,0 +1,4 @@
+package com.eerussianguy.firmalife.compat.crafttweaker;
+
+public class CTCracking {
+}
