+++ src/main/java/com/eerussianguy/firmalife/compat/jei/category/CastingCategoryFL.java
@@ -1,9 +1,9 @@
-package com.eerussianguy.firmalife.jei.category;
+package com.eerussianguy.firmalife.compat.jei.category;
 
 import net.minecraft.client.Minecraft;
 import net.minecraft.util.ResourceLocation;
 
-import com.eerussianguy.firmalife.jei.wrapper.CastingRecipeWrapperFL;
+import com.eerussianguy.firmalife.compat.jei.wrapper.CastingRecipeWrapperFL;
 import mezz.jei.api.IGuiHelper;
 import mezz.jei.api.gui.*;
 import mezz.jei.api.ingredients.IIngredients;
